Fix TypedProperties related to mouse control

+++ src/engine/playloop/p_lights.cc
@@ -686,7 +686,7 @@ typedef struct {
     float factor;
 } fadebright_t;
 
-extern IntProperty i_brightness;
+extern FloatProperty i_brightness;
 
 //
 // T_FadeInBrightness
